requirements.txt
requests
aiodns
aiohttp
azure-storage-blob
pandas
colorama
pathlib
datetime